Removed lxml and replication tool
The lxml package was causing issue due to it's system library dependency
Using minidom is standard package and does not require installing a
a system level package.

Due to no longer supporting fuzzy html parsing we also removed tree replication
tool. This was already removed from the AEM 6.1 version so it's not a big thing.
Must remember to add package replication to the package tool though.